#1354 Move GOConstLoopBoundsTrans tests to its own file
