+++ src/psyclone/tests/domain/gocean/goloop_test.py
@@ -90,49 +90,6 @@ def test_goloop_no_children():
         goiloop.gen_code(None)
 
 
-@pytest.mark.xfail(reason="Relocate to GOConstLoopBoundTrans?")
-def test_goloop_unsupp_offset():
-    ''' Attempt to generate code for a loop with constant bounds with
-    an unsupported index offset '''
-    gosched = GOInvokeSchedule('name', [])
-    gojloop = GOLoop(parent=gosched, loop_type="outer")
-    gosched.addchild(gojloop)
-    goiloop = GOLoop(parent=gojloop.loop_body, loop_type="inner")
-    gojloop.loop_body.addchild(goiloop)
-    gokern = GOKern(None)
-    # Set the index-offset of this kernel to a value that is not
-    # supported when using constant loop bounds
-    gokern._index_offset = "offset_se"
-    goiloop.loop_body.addchild(gokern)
-    with pytest.raises(GenerationError):
-        goiloop.gen_code(None)
-
-
-@pytest.mark.xfail(reason="Relocate to GOConstLoopBoundTrans?")
-def test_goloop_unmatched_offsets():
-    ''' Attempt to generate code for a loop with constant bounds with
-    two different index offsets '''
-    gosched = GOInvokeSchedule('name', [])
-    gojloop = GOLoop(parent=gosched, loop_type="outer")
-    gosched.addchild(gojloop)
-    goiloop = GOLoop(parent=gojloop.loop_body, loop_type="inner")
-    gojloop.loop_body.addchild(goiloop)
-    gokern1 = GOKern()
-    gokern2 = GOKern()
-    # Set the index-offset of this kernel to a value that is not
-    # supported when using constant loop bounds
-    gokern1._index_offset = "go_offset_ne"
-    gokern2._index_offset = "go_offset_sw"
-    goiloop.loop_body.addchild(gokern1)
-    goiloop.loop_body.addchild(gokern2)
-    with pytest.raises(GenerationError) as excinfo:
-        goiloop.gen_code(None)
-    # Note that the kernels do not have a name, so there are empty quotes
-    assert "All Kernels must expect the same grid offset but kernel '' " \
-        "has offset 'go_offset_sw' which does not match 'go_offset_ne'" \
-        in str(excinfo.value)
-
-
 def test_goloop_bounds_invalid_iteration_space():
     ''' Check that the _upper/lower_bound() methods raise the expected error
     if the iteration space is not recognised. '''
@@ -200,7 +157,11 @@ def test_goloop_validate_loop():
             in str(err.value))
 
     class GOKernMock(GOKern):
+        ''' Mock class of GOKern for this test'''
         def __init__(self):
+            ''' Overrided constructor to initialize it just as a
+            PSyIR node'''
+            # pylint: disable=super-init-not-called, non-parent-init-called
             Node.__init__(self)  # Ignore hierarchy constructors
 
     # Test Loop containing kernels with different offsets
